Preprocessing_genome and BS_Seekre_II are made to work with json instead of shelve due to problems with Berkeley DB.
Preprocessing_genome supports multiple genomes placed in one directory (reference_genome), which is located by default in the same directory as the python scripts (and not in the current working directory as before).

BS_Seeker_II requires a -g option that tells the name of the reference genome. It is looked up by this name in reference_genomes.

BS_Seeker handles correctly the case where the input file is given by its full or relative path.

Lots of code refactoring, too.